+++ noodles-vcf/src/record/genotypes/genotype/field.rs
@@ -27,12 +27,18 @@ pub enum ParseError {
     InvalidValue(value::ParseError),
 }
 
-impl error::Error for ParseError {}
+impl error::Error for ParseError {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        match self {
+            Self::InvalidValue(e) => Some(e),
+        }
+    }
+}
 
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::InvalidValue(e) => write!(f, "invalid value: {}", e),
+            Self::InvalidValue(_) => f.write_str("invalid value"),
         }
     }
 }