+++ noodles-vcf/src/record/genotypes/genotype/field/value.rs
@@ -125,7 +125,16 @@ pub enum ParseError {
     InvalidString(str::Utf8Error),
 }
 
-impl error::Error for ParseError {}
+impl error::Error for ParseError {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        match self {
+            Self::InvalidInteger(e) => Some(e),
+            Self::InvalidFloat(e) => Some(e),
+            Self::InvalidString(e) => Some(e),
+            _ => None,
+        }
+    }
+}
 
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {