+++ noodles-vcf/src/record/info/field.rs
@@ -121,15 +121,23 @@ pub enum ParseError {
     InvalidValue(value::ParseError),
 }
 
-impl error::Error for ParseError {}
+impl error::Error for ParseError {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        match self {
+            Self::InvalidKey(e) => Some(e),
+            Self::InvalidValue(e) => Some(e),
+            _ => None,
+        }
+    }
+}
 
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::MissingKey => f.write_str("missing key"),
-            Self::InvalidKey(e) => write!(f, "invalid key: {}", e),
+            Self::InvalidKey(_) => f.write_str("invalid key"),
             Self::MissingValue => f.write_str("missing value"),
-            Self::InvalidValue(e) => write!(f, "invalid value: {}", e),
+            Self::InvalidValue(_) => f.write_str("invalid value"),
         }
     }
 }