+++ noodles-vcf/src/record/quality_score.rs
@@ -54,14 +54,22 @@ pub enum ParseError {
     InvalidValue(TryFromFloatError),
 }
 
-impl error::Error for ParseError {}
+impl error::Error for ParseError {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        match self {
+            Self::Empty => None,
+            Self::Invalid(e) => Some(e),
+            Self::InvalidValue(e) => Some(e),
+        }
+    }
+}
 
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Empty => f.write_str("empty input"),
-            Self::Invalid(e) => write!(f, "invalid input: {}", e),
-            Self::InvalidValue(e) => write!(f, "invalid value: {}", e),
+            Self::Invalid(_) => f.write_str("invalid input"),
+            Self::InvalidValue(_) => f.write_str("invalid value"),
         }
     }
 }