+++ noodles-vcf/src/record/reference_bases.rs
@@ -52,14 +52,21 @@ pub enum ParseError {
     InvalidBase(base::TryFromCharError),
 }
 
-impl error::Error for ParseError {}
+impl error::Error for ParseError {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        match self {
+            Self::InvalidBase(e) => Some(e),
+            _ => None,
+        }
+    }
+}
 
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Empty => f.write_str("empty input"),
             Self::Missing => f.write_str("missing input"),
-            Self::InvalidBase(e) => write!(f, "invalid base: {}", e),
+            Self::InvalidBase(_) => f.write_str("invalid base"),
         }
     }
 }